Read detail below:
- Add id field (with auto increasement) on User and Card Model
- Add getter and setter to each model
- Change id datatype from "Integer" -> "Long"
- Modify JpaRepository for each Repo

+++ WG26/src/main/java/tdtu/edu/un/WG26/Repository/AppRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import tdtu.edu.un.WG26.Model.App;
 
 @Repository
-public interface AppRepository extends JpaRepository<App, Integer>{
+public interface AppRepository extends JpaRepository<App, Long>{
 	public App findByAppName(String appName);
 }
